requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-swagger==1.0.16
aiosignal==1.2.0
appnope==0.1.2
argon2-cffi==21.1.0
arrow==1.2.1
async-lru==1.0.2
async-timeout==4.0.1
asyncssh==2.8.1
attrs==21.2.0
backcall==0.2.0
backports.ssl==0.0.9
binaryornot==0.4.4
black==21.11b1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cookiecutter==1.7.3
cryptography==36.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
distro==1.6.0
entrypoints==0.3
frozenlist==1.2.0
future==0.18.2
genie==21.10
genie.libs.clean==21.10
genie.libs.conf==21.10
genie.libs.filetransferutils==21.10
genie.libs.health==21.10
genie.libs.ops==21.10
genie.libs.parser==21.10
genie.libs.robot==21.10
genie.libs.sdk==21.10
genie.telemetry==21.10
genie.trafficgen==21.10
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
IxNetwork==9.0.1915.16
ixnetwork-restpy==1.1.3
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jsonpickle==2.0.0
jsonschema==4.2.1
junit-xml==1.9
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
notebook==6.4.6
ntc-templates==2.3.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
poyo==0.5.0
prettytable==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psutil==5.8.0
ptyprocess==0.7.0
pyats==21.10
pyats.aereport==21.10
pyats.aetest==21.10
pyats.async==21.10
pyats.connections==21.10
pyats.contrib==21.10.1
pyats.datastructures==21.10
pyats.easypy==21.10
pyats.kleenex==21.10
pyats.log==21.10
pyats.reporter==21.10.1
pyats.results==21.10
pyats.robot==21.10
pyats.tcl==21.10
pyats.topology==21.10
pyats.utils==21.10
pycparser==2.21
pyftpdlib==1.5.6
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==3.14.2
python-slugify==5.0.2
python-socketio==4.6.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
regex==2021.11.10
requests==2.26.0
requests-toolbelt==0.9.1
rich==10.15.2
robotframework==4.1.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scrapli==2021.7.30
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
textfsm==1.1.2
tftpy==0.8.0
tokenize-rt==4.2.1
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
unicon==21.10
unicon.plugins==21.10
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
xmltodict==0.12.0
yamllint==1.26.3
yarl==1.7.2
zipp==3.6.0
